Move metrics to seperate module
