+++ packages/beacon-node/src/chain/archiveStore/historicalState/types.ts
@@ -1,8 +1,6 @@
 import {ModuleThread} from "@chainsafe/threads";
 import {BeaconConfig, SpecJson} from "@lodestar/config";
 import {LoggerNode, LoggerNodeOpts} from "@lodestar/logger/node";
-import {BeaconStateTransitionMetrics} from "@lodestar/state-transition";
-import {Gauge, Histogram} from "@lodestar/utils";
 import {Metrics} from "../../../metrics/index.js";
 
 export type HistoricalStateRegenInitModules = {
@@ -41,14 +39,3 @@ export enum RegenErrorType {
   invalidStateRoot = "invalid_state_root",
   blockProcessing = "block_processing",
 }
-
-export type HistoricalStateRegenMetrics = BeaconStateTransitionMetrics & {
-  regenTime: Histogram;
-  loadStateTime: Histogram;
-  stateTransitionTime: Histogram;
-  stateTransitionBlocks: Histogram;
-  stateSerializationTime: Histogram;
-  regenRequestCount: Gauge;
-  regenSuccessCount: Gauge;
-  regenErrorCount: Gauge<{reason: RegenErrorType}>;
-};
